refactor: move types and constants into separate files
