+++ src/App/components/Header/types.ts
@@ -1,0 +1,1 @@
+export type NavigateFunction = () => void;
