+++ src/App/components/HeaderNav/types.ts
@@ -1,0 +1,4 @@
+export interface MenuItem {
+    label: string;
+    route: string;
+}
