+++ src/App/components/InfoCard/InfoCard.tsx
@@ -3,33 +3,23 @@ import styles from './Card.module.scss';
 import Text from 'components/Text/Text';
 import { useNavigate } from 'react-router';
 import { routes } from "config/routes.config";
+import type { CardProps } from './types'
 
-export type CardProps = {
-    className?: string,
-    image: string;
-    captionSlot?: React.ReactNode;
-    title: React.ReactNode;
-    subtitle: React.ReactNode;
-    contentSlot?: React.ReactNode;
-    onClick?: React.MouseEventHandler;
-    actionSlot?: React.ReactNode;
-    itemDocumentId: string;
-};
 
 const InfoCard: React.FC<CardProps> = ({ className, image, captionSlot, title, subtitle, contentSlot, onClick, actionSlot, itemDocumentId }) => {
-    const actualClassName = `${styles.card} ${className? className : ''}`;
+    const actualClassName = `${styles.card} ${className ? className : ''}`;
     const navigate = useNavigate();
 
     return (
-        <div className={actualClassName} onClick={onClick? onClick : () => navigate(routes.recipe.create(itemDocumentId))}> 
+        <div className={actualClassName} onClick={onClick ? onClick : () => navigate(routes.recipe.create(itemDocumentId))}>
             <div className={styles['card__image-container']}>
-            <img src={image} alt='картинка' className={styles['card__image']} />
+                <img src={image} alt='картинка' className={styles['card__image']} />
             </div>
             <div className={styles['card__info']}>
                 <div className={styles['card__description']}>
                     {captionSlot && <p className={styles['card__caption']}>{captionSlot}</p>}
                     <div className={styles['card__title']}>
-                    <Text className="pb-8" view={'p-20'} weight='medium' maxLines={1} color='primary'>{title}</Text>
+                        <Text className="pb-8" view={'p-20'} weight='medium' maxLines={1} color='primary'>{title}</Text>
                     </div>
                     <Text className="pb-8" view={'p-16'} weight='normal' maxLines={2} color='secondary'>{subtitle}</Text>
                 </div>
@@ -47,4 +37,4 @@ const InfoCard: React.FC<CardProps> = ({ className, image, captionSlot, title, s
     )
 };
 
-export default memo(InfoCard) ;
+export default memo(InfoCard);
