+++ src/App/components/InfoCard/types.ts
@@ -1,0 +1,11 @@
+export type CardProps = {
+    className?: string,
+    image: string;
+    captionSlot?: React.ReactNode;
+    title: React.ReactNode;
+    subtitle: React.ReactNode;
+    contentSlot?: React.ReactNode;
+    onClick?: React.MouseEventHandler;
+    actionSlot?: React.ReactNode;
+    itemDocumentId: string;
+};
