+++ src/App/components/Input/Input.tsx
@@ -1,14 +1,7 @@
 import React, { memo } from 'react';
 import styles from './Input.module.scss';
+import type { InputProps } from './types';
 
-export type InputProps = Omit<
-  React.InputHTMLAttributes<HTMLInputElement>,
-  'onChange' | 'value'
-> & {
-  value: string;
-  onChange: (value: string) => void;
-  afterSlot?: React.ReactNode;
-};
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
   ({ value, onChange, afterSlot, ...rest }, ref) => {
