+++ src/App/components/Input/types.ts
@@ -1,0 +1,8 @@
+export type InputProps = Omit<
+    React.InputHTMLAttributes<HTMLInputElement>,
+    'onChange' | 'value'
+> & {
+    value: string;
+    onChange: (value: string) => void;
+    afterSlot?: React.ReactNode;
+};
