+++ src/App/components/MultiDropdown/index.ts
@@ -1,2 +1,3 @@
 export { default } from './MultiDropdown';
 export * from './MultiDropdown';
+export type { Option } from './types';
