+++ src/App/components/MultiDropdown/types.ts
@@ -1,0 +1,13 @@
+export type Option = {
+    key: string;
+    value: string;
+};
+
+export type MultiDropdownProps = {
+    className?: string;
+    options: Option[];
+    value: Option[];
+    onChange: (value: Option[]) => void;
+    disabled?: boolean;
+    getTitle: (value: Option[]) => string;
+};
