+++ src/App/components/Pagination/Pagination.tsx
@@ -2,19 +2,9 @@ import styles from './Pagination.module.scss';
 import arrowBackIcon from 'assets/images/arrow-right.svg'
 import arrowForwardIcon from 'assets/images/arrow-rightSingle.svg'
 import { useSearchParams } from 'react-router';
+import type { PaginationProps } from './types';
+import { getNumberCountArr } from 'utils/helpers';
 
-interface PaginationProps {
-    pageCount: number;
-    actualPage: number;
-}
-
-export const getNumberCountArr = (pageCount: number): number[] => {
-    const result = [];
-    for (let i = 1; i <= pageCount; i += 1) {
-        result.push(i);
-    }
-    return result
-}
 
 const Pagination = ({ pageCount, actualPage }: PaginationProps) => {
     const [searchParams, setSearchParams] = useSearchParams();
