+++ src/App/components/Pagination/types.ts
@@ -1,0 +1,4 @@
+export interface PaginationProps {
+    pageCount: number;
+    actualPage: number;
+}
