+++ src/App/pages/CatalogPage/index.ts
@@ -1,1 +1,2 @@
 export { default } from './CatalogPage';
+export type { Ingredient, Recipe } from './types';
