+++ src/App/pages/CatalogPage/types.ts
@@ -1,0 +1,48 @@
+export interface Image {
+    id: number;
+    documentId: string;
+    name: string;
+    alternativeText: string | null;
+    caption: string | null;
+    url: string
+}
+
+export interface Ingredient {
+    id: number;
+    name: string;
+    amount: number;
+    unit: string | null;
+}
+
+export type Equipment = {
+    id: number,
+    name: string
+}
+
+export type Direction = {
+    description: string,
+    id: number,
+    image: null | string
+}
+
+export interface Recipe {
+    calories: number;
+    cookingTime: number;
+    createdAt: string;
+    documentId: string;
+    id: number;
+    images: Image[];
+    ingradients: Ingredient[];
+    likes: number;
+    name: string;
+    preparationTime: number;
+    publishedAt: string;
+    rating: number;
+    servings: number;
+    summary: string;
+    totalTime: number;
+    updatedAt: string;
+    vegetarian: boolean;
+    equipments?: Equipment[];
+    directions?: Direction[];
+}
