+++ src/components/Button/types.ts
@@ -1,0 +1,5 @@
+export type ButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
+    loading?: boolean;
+    children: React.ReactNode;
+    'data-testid'?: string;
+};
