+++ src/components/Loader/config.ts
@@ -1,0 +1,5 @@
+export const loaderSize = {
+    s: '24px',
+    m: '48px',
+    l: '60px',
+};
