+++ src/components/Loader/types.ts
@@ -1,0 +1,4 @@
+export type LoaderProps = {
+    size?: 's' | 'm' | 'l';
+    className?: string;
+};
