+++ src/components/Text/Text.tsx
@@ -1,15 +1,7 @@
 import * as React from 'react';
 import styles from './Text.module.scss';
+import type { TextProps } from './types';
 
-export type TextProps = {
-    className?: string;
-    view?: 'title' | 'button' | 'p-20' | 'p-18' | 'p-16' | 'p-14';
-    tag?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'div' | 'p' | 'span';
-    weight?: 'normal' | 'medium' | 'bold';
-    children: React.ReactNode;
-    color?: 'primary' | 'secondary' | 'accent';
-    maxLines?: number;
-};
 
 const Text: React.FC<TextProps> = ({
     className, view, tag = 'p', weight, children, color, maxLines = 100
