+++ src/components/Text/types.ts
@@ -1,0 +1,9 @@
+export type TextProps = {
+    className?: string;
+    view?: 'title' | 'button' | 'p-20' | 'p-18' | 'p-16' | 'p-14';
+    tag?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'div' | 'p' | 'span';
+    weight?: 'normal' | 'medium' | 'bold';
+    children: React.ReactNode;
+    color?: 'primary' | 'secondary' | 'accent';
+    maxLines?: number;
+};
