+++ src/components/icons/CheckIcon/CheckIcon.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { IconProps } from '../Icon';
+import type { IconProps } from './../Icon';
 
 const CheckIcon: React.FC<IconProps> = (props) => {
     const getClassNames = () => {
