+++ src/components/icons/Icon/index.ts
@@ -1,1 +1,2 @@
-export * from './Icon';
+export * from './types';
+export type { IconProps } from './types';
