+++ tsconfig.app.json
@@ -42,6 +42,9 @@
       "App": [
         "App/*"
       ],
+      "utils": [
+        "utils/*"
+      ],
     }
   },
   "include": [
