+++ tsconfig.json
@@ -1,19 +1,37 @@
 {
   "files": [],
   "references": [
-    { "path": "./tsconfig.app.json" },
-    { "path": "./tsconfig.node.json" }
+    {
+      "path": "./tsconfig.app.json"
+    },
+    {
+      "path": "./tsconfig.node.json"
+    }
   ],
   "compilerOptions": {
     "baseUrl": "src",
     "paths": {
-      "components": ["components/*"],
-      "config": ["config/*"],
-      "styles": ["styles/*"],
-      "utils": ["utils/*"],
-      "assets": ["assets/*"],
-      "App": ["App/*"]
-
+      "components": [
+        "components/*"
+      ],
+      "config": [
+        "config/*"
+      ],
+      "styles": [
+        "styles/*"
+      ],
+      "utils": [
+        "utils/*"
+      ],
+      "assets": [
+        "assets/*"
+      ],
+      "App": [
+        "App/*"
+      ],
+      "utils": [
+        "utils/*"
+      ],
     }
   }
 }
